+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.22.0
 
 - New `Feature.getBackground()` method, which returns an object containing the test steps of the background of the features
+- New `Examples.getTags()` and `Example.getTags()` methods to give back the tags from examples
 - New `HttpClient.multipart(...)` and `RestClient.multipart(...)` methods to create a multipart/form-data request
 - The `RestClient.sendAndReceive(...)` method can receive response with `byte[]` or `InputStream` type
 - The test result XML contains the category of the test steps